+++ src/utils/writeClientClients.ts
@@ -3,7 +3,7 @@ import { relative, resolve } from 'path';
 import { writeFile } from './fileSystem.js';
 import { formatCode as f } from './formatCode.js';
 import { formatIndentation as i } from './formatIndentation.js';
-import { WriteClientPartContext } from './writeClientTypes';
+import { WriteClientPartContext } from './types.js';
 
 /**
  * Generate Client using the Handlebar template and write to disk.
